Extract setting danger-js path as a fixture
